requirements.txt
beautifulsoup4==4.12.3
chromadb==0.5.13
fastapi==0.115.2
openai==1.51.2
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
tiktoken==0.8.0
uvicorn==0.31.1
